requirements.txt
# Python dependencies for optional tooling.
numpy==1.26.4
matplotlib==3.9.2
# Pillow is required for image generation.
# Use any modern version; if PyPI is unreachable, install via system
# package (e.g., `sudo apt-get install python3-pil`) or a downloaded wheel.
Pillow>=10.0

# Add your other deps below, pinned if possible:
numpy==1.26.4
matplotlib==3.9.2
# fastapi==0.114.0
# uvicorn==0.30.6
